ai_agent_job_search/requirements.txt
langchain
langchain-openai
tavily-python
pandas
python-dotenv
streamlit
